Hacky version works! Now I "just" need to parse .gv.
So, this does not use reflection or anything nice like that, because
afaik, it's not there yet. But it does work for the ATM example, and I
believe it should work for more complex examples as well (TODO: come up
with these).

It's somewhat halfway between "draw the rest of the owl" and my Idris2
Petri Net implementation. Perhaps I could draw on the PN stuff in terms
of the soundness of the DSL?