Added handling of EdgeStmts. Needs to be subset of DOT
The current implementation works, but it is too close to DSA when DOTDSA
is meant to be a subset of DOT. Case in point being the functions that
return a `Maybe (List _)` -- these convert DOT to something closer to a
DSA, which is fine in theory, but in practice the DOTDSA should be
easily convertible to DOT, i.e. the step of separating things out into
lists is too early at this stage.
Need to refactor.