Redid `Label` loser to the idea of DOTDSA (I think)
Label is now a lot closer to what I believe was my original idea of
DOTDSA: A restricted subset of DOT for describing sensible DSAs.
Moving on to edges. That should show whether I'm now finally headed down
the right path.

Cleaning this code up is going to be a pain...