Almost there. Need to fix the _subtlest_ of bugs!
A massive thank you to @gallais for spotting an error that almost had me
convinced I was going insane:
`v="InsertCard"` means that `v` _contains_ those quotes if you're
displaying it using `++` and not `show` (which I was). This led to the
fun bug of:
```idris
v="InsertCard"
labelValOK v: False

> labelValOK "InsertCard"
True
```
which, understandably, had me doubting everything I was seeing.
I need to fix this, but not tonight.

Also, code needs cleaning up after all the debug `export` statements and
`IO ()` stuff I threw in there to try to get a better understanding of
what on Earth was happening...