Fix various bugs. DOTDSA NOW WORKS!!!
First and foremost, this fixes the bug with the extra quotes in a
`StringID` by simply taking the substring from the second to the
penultimate character of the string.

Some other bugs popped up, e.g. `isDepEdge` had problems with
parentheses. I tried being clever using `split`, but the simple `span` +
`SnocList` solution just looks nice and works fine.

Having fixed these things, it seems to now work correctly for the ATM
example, which is really neat! ^^
Now for cleanup...