Code cleanup: Remove excessive DOTDSA implementation
The lines left in `DSAGen/DOTDSA.idr` actually represent the idea I had
for DOTDSA: A restricted subset of DOT that it is easier to convert to
Idris.

Whether it is actually easier to convert is the next question, but at
least the current implementation nicely translates back to DOT, which
was the goal.