Avoid name clashes with `DSAGen/DSL.idr`
Turns out I'd named my DOTDSA types exactly the same as the ones used in
the DSL for DSAs; obviously these are good names, but the clashes are
not ideal ^^;

Renamed the DOTDSA types to avoid the clashes. Probably should have
spotted this during the cleanup, but oh well. I believe things are now
ready to get going towards implementing a `toDSA` function!