Bit of refactoring; started drafting "universal edges"
Some of the functions I defined in `DOTDSA` as part of its `toDSA`
implementation fit much better in the `DSL` file.

I have an idea for how to detect universal edges, but I haven't quite
figured out how to implement it yet. Might need to refine the DSL first,
if I want to reduce the need for `idris_crash` things.