Refined the DSL in a separate file. It works!
I decided to improve the DSL a bit, just to see if there was any reason
I was using `assert_total $ idris_crash` in the original `DSL` or if I
could get rid of those. It seems the answer is "Yes"!

Now I need to implement `toDSA2` for `DOTDSA`, but then (if that works
fine) I will update the DSL to the refined one.