DOTDSA successfully mapped to refined DSL (DSL2)
I think that was slightly less painful than the mapping from DOTDSA to
the original DSL. There are possibly some refinements to be done in the
types, e.g. indexing the labels by their number of values(?), but that
is a step for later. For now, I'm just happy this mapping worked.

Once again, lots of code cleanup to do...