[ WIP ] Initial new implementation of parser
This now half parses things as expected: Some commands work, but the
combinations don't. This is probably due to how parser-alternatives
work in combination with `commit`. I'll likely have to combine a couple
of parsers.

Currently, data constructors can also not be nested due to the
parentheses being consumed, leading to rejection of strings like
"Send:(Ack)"
in favour of
"Send:(Ack))"
...