[ WIP ] Comment out unused code; Re-implement `toDOTDSA`
It was simpler to just re-implement the function than decipher the
"non-covering case of" error messages. Which it also turned out were
wrong (more investigation on the desugaring of `let a = foo | b =>`
needed; see discord).

The code that's been commented out using `----` _should_ be unnecessary,
but we'll see if it actually is or if it needs reimplementing/swapping
with different code. When all is well, it will be deleted.
I've also introduced a hole (in the shape of an entire function) to
potentially re-implement... ^^;;