[ WIP ] Initial draft of the new DSL
Having tried to integrate the new syntax+structures into the old code,
it seems clear that the DSL needs expanding to account for the new types
of commands/edges. I think the easiest solution to this is to simply use
those types directly in a new DSL, along with some proofs.

This commit introduces an initial draft of that new DSL. The next parts
will be to try and integrate the lexing+parsing with it, which may be
more or less successful. Then there's the Idris codegen to do as well...