Use `Subset` and `Split` directly in DSA type
I was passing auto-implicit arguments explicitly in places where I
expected them to just work (tm). This usually suggests something should
be changed with the type. In this case, all I really care about is that
the proofs are Rig0 (i.e. erased at runtime), so using `Subset` and
`Split` internally in `DSAv2` is fine (and those are what I'm
constructing anyway).