Start on DOT edge -> DSA command conversion
This is slightly more complicated than it initially appears since labels
may contain multiple commands separated by ';', and so we need to handle
each of these, returning a `List DSALabel` rather than simply one
`DSALabel` per label.

This also means that the checking of the `Assign` in the state
generation was wrong. It's now a simpler check, but that's okay since
we'll catch the other stuff when generating edges.

I need to:
  * check that the `EdgeStmt` is well-formed (just for sanity's sake),
  * handle the converting each of the commands (`traverse`?),
  * figure out how to flatten the `List (List DSALabel)` into a
    `List DSALabel`,
  * deal with Universal Edge detection (ugh...)