Start cg and proofs for advanced edges
These need separate handling, so the functions will restrict them as
appropriate (we can satisfy the relevant `auto`s by pattern-matching on
the `naws` in the `Split`).

Any edge involving dependent stuff will need accumulating, in order to
know all the possible values the result the edge depends on can take.
This will be some work...