Draft genNonPlainNonDepEdge; fix magic ? in nawsSubset
Thanks to Guillaume (gallais) for helping me understand what the `?` was
doing and why `(reverse nawsSubset)` wasn't working as I thought it
would. Since we unfortunately can't refer to things inside
`where`-blocks and the `Split` manipulations are quite involved at this
point, the easiest solution is just to avoid implicit generalisation by
having the name start with an upper-case letter.