[ dsa-gen ] Rely on `idris2` instead of `contrib`
The contrib parsing libraries have changed (or maybe they were removed,
I can't recall) and it is better to use the internal libraries anyhow,
as these are required for Idris2 itself to work. Thankfully porting the
code wasn't too much effort. Mainly needed to update a couple of type
annotations to include warnings, and then update the pattern matches
that suddenly were lacking an argument.